refactor: [exporter interface] Refactor exporter to an interface
