+++ server/ingester/flow_log/exporter/config_test.yaml
@@ -1,0 +1,17 @@
+ingester:
+  exporters:
+    - name: test exporter
+      exporter_type: otlp-exporter
+      otlp-exporter:
+        enabled: false
+        addr: 127.0.0.1:4317
+        export-datas: [cbpf-net-span,ebpf-sys-span]
+        export-data-types: [ service_info,tracing_info,network_layer,flow_info,transport_layer,application_layer,metrics ]
+        queue-count: 4
+        queue-size: 100000
+        export-custom-k8s-labels-regexp:
+        export-only-with-traceid: false
+        export-batch-count: 32
+        grpc-headers:
+          key1: value1
+          key2: value2
