+++ server/ingester/flow_log/exporter/otlp_exporter/util.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package exporter
+package otlp_exporter
 
 import (
 	"bytes"
